Reorganizing module structure.
